fix: 消除 compileKotlin 编译警告（废弃覆盖/API调用/多余转换）

+++ core/src/main/kotlin/ic2_120/content/block/energy/EnergyNetwork.kt
@@ -337,7 +337,7 @@ class EnergyNetwork : SnapshotParticipant<EnergyNetwork.NetworkSnapshot>() {
             log.debug("[导线烧毁] 电网输出等级=$outputLevel，可烧毁导线数量=${toBurn.size}（电压等级 < $outputLevel）")
         }
 
-        val burnCount = (toBurn.size * underTierCableBurnRatio).toInt().coerceIn(0, toBurn.size)
+        val burnCount = (toBurn.size * underTierCableBurnRatio).coerceIn(0, toBurn.size)
         if (burnCount <= 0) return
 
         if (ENABLE_CABLE_BURN_LOG) {
